refactor: split up container handle from bootstrapper feat: make bootstrapper not implement IMixinConfigSource
